--- BE/network/client_data/test.txt
@@ -1,1 +1,0 @@
-Hello world
