+++ Client_TCP/FE/config.js
@@ -1,0 +1,1 @@
+var BASE_URL='http://127.0.0.1:5000/';
