+++ Client_TCP/FE/home.html
@@ -26,14 +26,17 @@
 
   <link href="https://getbootstrap.com/docs/5.3/assets/css/docs.css" rel="stylesheet">
   <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"></script>
-  <script src="js/api/api.js"></script>
 </head>
 
 <body>
   <div class="sidebar">
     <div class="row mr-0">
-      <div class="col-2 col-sidebar pl-4" style="height: 100vh; ">
-        <?php include "sidebar.php" ?>
+      <div class="col-2 col-sidebar pl-4 includedContent" style="height: 100vh;">
+        <script>
+          $(function () {
+            $(".includedContent").load("sidebar.html");
+          });
+        </script>
       </div>
 
 
@@ -78,91 +81,104 @@
         <div class="list-team">
           <div>
             <div class="row div-list-team">
-              <a href="team.php" class="col-2 team" style="margin-bottom: 34px; margin-right: 16px; margin-left: 18px;">
+              <a href="team.html" class="col-2 team"
+                style="margin-bottom: 34px; margin-right: 16px; margin-left: 18px;">
                 <div class="team-element">
                   <div class="img-div"></div>
                 </div>
                 <br>
                 <h6>Network Programing</h6>
               </a>
-              <a href="team.php" class="col-2 team" style="margin-bottom: 34px; margin-right: 16px; margin-left: 18px;">
+              <a href="team.html" class="col-2 team"
+                style="margin-bottom: 34px; margin-right: 16px; margin-left: 18px;">
                 <div class="team-element">
                   <div class="img-div"></div>
                 </div>
                 <br>
                 <h6>Viet Nhat</h6>
               </a>
-              <a href="team.php" class="col-2 team" style="margin-bottom: 34px; margin-right: 16px; margin-left: 18px;">
+              <a href="team.html" class="col-2 team"
+                style="margin-bottom: 34px; margin-right: 16px; margin-left: 18px;">
                 <div class="team-element">
                   <div class="img-div"></div>
                 </div>
                 <br>
                 <h6>Tran Hong Quan</h6>
               </a>
-              <a href="team.php" class="col-2 team" style="margin-bottom: 34px; margin-right: 16px; margin-left: 18px;">
+              <a href="team.html" class="col-2 team"
+                style="margin-bottom: 34px; margin-right: 16px; margin-left: 18px;">
                 <div class="team-element">
                   <div class="img-div"></div>
                 </div>
                 <br>
                 <h6>Hoang Minh Nguyet</h6>
               </a>
-              <a href="team.php" class="col-2 team" style="margin-bottom: 34px; margin-right: 16px; margin-left: 18px;">
+              <a href="team.html" class="col-2 team"
+                style="margin-bottom: 34px; margin-right: 16px; margin-left: 18px;">
                 <div class="team-element">
                   <div class="img-div"></div>
                 </div>
                 <br>
                 <h6>Nguyen Hai Nam</h6>
               </a>
-              <a href="team.php" class="col-2 team" style="margin-bottom: 34px; margin-right: 16px; margin-left: 18px;">
+              <a href="team.html" class="col-2 team"
+                style="margin-bottom: 34px; margin-right: 16px; margin-left: 18px;">
                 <div class="team-element">
                   <div class="img-div"></div>
                 </div>
                 <br>
                 <h6>Team A</h6>
               </a>
-              <a href="team.php" class="col-2 team" style="margin-bottom: 34px; margin-right: 16px; margin-left: 18px;">
+              <a href="team.html" class="col-2 team"
+                style="margin-bottom: 34px; margin-right: 16px; margin-left: 18px;">
                 <div class="team-element">
                   <div class="img-div"></div>
                 </div>
                 <br>
                 <h6>Team B</h6>
               </a>
-              <a href="team.php" class="col-2 team" style="margin-bottom: 34px; margin-right: 16px; margin-left: 18px;">
+              <a href="team.html" class="col-2 team"
+                style="margin-bottom: 34px; margin-right: 16px; margin-left: 18px;">
                 <div class="team-element">
                   <div class="img-div"></div>
                 </div>
                 <br>
                 <h6>Team C</h6>
               </a>
-              <a href="team.php" class="col-2 team" style="margin-bottom: 34px; margin-right: 16px; margin-left: 18px;">
+              <a href="team.html" class="col-2 team"
+                style="margin-bottom: 34px; margin-right: 16px; margin-left: 18px;">
                 <div class="team-element">
                   <div class="img-div"></div>
                 </div>
                 <br>
                 <h6>Team D</h6>
               </a>
-              <a href="team.php" class="col-2 team" style="margin-bottom: 34px; margin-right: 16px; margin-left: 18px;">
+              <a href="team.html" class="col-2 team"
+                style="margin-bottom: 34px; margin-right: 16px; margin-left: 18px;">
                 <div class="team-element">
                   <div class="img-div"></div>
                 </div>
                 <br>
                 <h6>Network Programing</h6>
               </a>
-              <a href="team.php" class="col-2 team" style="margin-bottom: 34px; margin-right: 16px; margin-left: 18px;">
+              <a href="team.html" class="col-2 team"
+                style="margin-bottom: 34px; margin-right: 16px; margin-left: 18px;">
                 <div class="team-element">
                   <div class="img-div"></div>
                 </div>
                 <br>
                 <h6>Network Programing</h6>
               </a>
-              <a href="team.php" class="col-2 team" style="margin-bottom: 34px; margin-right: 16px; margin-left: 18px;">
+              <a href="team.html" class="col-2 team"
+                style="margin-bottom: 34px; margin-right: 16px; margin-left: 18px;">
                 <div class="team-element">
                   <div class="img-div"></div>
                 </div>
                 <br>
                 <h6>Network Programing</h6>
               </a>
-              <a href="team.php" class="col-2 team" style="margin-bottom: 34px; margin-right: 16px; margin-left: 18px;">
+              <a href="team.html" class="col-2 team"
+                style="margin-bottom: 34px; margin-right: 16px; margin-left: 18px;">
                 <div class="team-element">
                   <div class="img-div"></div>
                 </div>
@@ -203,7 +219,8 @@
           <input type="text" id="create-team-des" placeholder="Description" style="margin-bottom: 3px;">
           <p class="p-message" id="p-message-create-team">The team name already exists</p>
           <button class="btn btn-success next-btn create-team-btn"
-            style="margin-left: auto; background-color: #4D4D4D; margin-top: 18px;" id="next-btn-create-team">Next</button>
+            style="margin-left: auto; background-color: #4D4D4D; margin-top: 18px;"
+            id="next-btn-create-team">Next</button>
         </div>
