+++ Client_TCP/FE/templates/login.css
@@ -110,7 +110,6 @@ body {
     top: 0;
     width: 900px;
     height: 100%;
-    background-image: url("ext.jpg");
     opacity: .8;
     background-size: cover;
     -webkit-transition: -webkit-transform 1.2s ease-in-out;
