--- FE/.env
@@ -1,1 +1,0 @@
-BASE_URL=http://127.0.0.1:5000/
